Extract common initialization into separate file
